Move get_available_song_ids_from_files into module
